reinstate example and run a short one during testing
